Move modules and add some type safety
