+++ src/parser/core/modules/index.ts
@@ -1,7 +1,9 @@
 import {About} from './About'
 import {ActionTimeline} from './ActionTimeline'
 import {Actors} from './Actors'
-import {AlwaysBeCasting} from './AlwaysBeCasting'
+import {AlwaysBeCasting} from './AlwaysBeCasting/AlwaysBeCasting'
+import {Interrupts} from './AlwaysBeCasting/Interrupts'
+import {Weaving} from './AlwaysBeCasting/Weaving'
 import {BrokenLog} from './BrokenLog'
 import {CastTime} from './CastTime'
 import {ChangeLog} from './ChangeLog'
@@ -16,7 +18,6 @@ import {Dummy} from './Dummy'
 import {EventsView} from './EventsView'
 import {GlobalCooldown} from './GlobalCooldown'
 import {Hijinks} from './Hijinks'
-import {Interrupts} from './Interrupts'
 import {InvalidEvent} from './InvalidEvent'
 import {Invulnerability} from './Invulnerability'
 import {Medicated} from './Medicated'
@@ -29,7 +30,6 @@ import {StatusTimeline} from './StatusTimeline'
 import {Suggestions} from './Suggestions'
 import {Timeline} from './Timeline'
 import {UnableToAct} from './UnableToAct'
-import {Weaving} from './Weaving'
 
 export const modules = [
 	About,
