+++ src/parser/jobs/blu/modules/AlwaysBeCasting.tsx
@@ -6,7 +6,7 @@ import {filter, oneOf} from 'parser/core/filter'
 import {dependency} from 'parser/core/Injectable'
 import {History} from 'parser/core/modules/ActionWindow/History'
 import {Actors} from 'parser/core/modules/Actors'
-import {AlwaysBeCasting as CoreAlwaysBeCasting} from 'parser/core/modules/AlwaysBeCasting'
+import {AlwaysBeCasting as CoreAlwaysBeCasting} from 'parser/core/modules/AlwaysBeCasting/AlwaysBeCasting'
 import {SimpleStatistic, Statistics} from 'parser/core/modules/Statistics'
 import {Suggestions, TieredSuggestion, SEVERITY} from 'parser/core/modules/Suggestions'
