+++ src/parser/jobs/blu/modules/Interrupts.tsx
@@ -1,6 +1,6 @@
 import {Trans} from '@lingui/react/macro'
 import {DataLink} from 'components/ui/DbLink'
-import {Interrupts as CoreInterrupts} from 'parser/core/modules/Interrupts'
+import {Interrupts as CoreInterrupts} from 'parser/core/modules/AlwaysBeCasting/Interrupts'
 
 export class Interrupts extends CoreInterrupts {
 	override suggestionContent = <Trans id="blu.interrupts.suggestion.content">
