+++ src/parser/jobs/blu/modules/Weaving.tsx
@@ -6,7 +6,7 @@ import {filter} from 'parser/core/filter'
 import {dependency} from 'parser/core/Injectable'
 import {Actors} from 'parser/core/modules/Actors'
 import {TieredSuggestion, SEVERITY} from 'parser/core/modules/Suggestions'
-import {Weaving, Weave} from 'parser/core/modules/Weaving'
+import {Weaving, Weave} from 'parser/core/modules/AlwaysBeCasting/Weaving'
 
 const MAX_CAST_TIME_FOR_SINGLE_WEAVE = 1500
 const SINGLE_WEAVE_MS = 1000
