+++ src/parser/jobs/drg/modules/Weaving.ts
@@ -1,5 +1,5 @@
 import {ActionKey} from 'data/ACTIONS'
-import {Weaving as CoreWeaving, Weave} from 'parser/core/modules/Weaving'
+import {Weaving as CoreWeaving, Weave} from 'parser/core/modules/AlwaysBeCasting/Weaving'
 
 // With the reduced animation lock, it's just stardiver that's the bad weave
 const JUMPS: ActionKey[] = [
