+++ src/parser/jobs/sch/modules/Weaving.tsx
@@ -1,7 +1,7 @@
 import {Trans} from '@lingui/react/macro'
 import {DataLink} from 'components/ui/DbLink'
 import {SEVERITY} from 'parser/core/modules/Suggestions'
-import {Weaving as CoreWeaving} from 'parser/core/modules/Weaving'
+import {Weaving as CoreWeaving} from 'parser/core/modules/AlwaysBeCasting/Weaving'
 
 const WEAVING_SEVERITY = {
 	1: SEVERITY.MINOR,
