+++ src/parser/jobs/sge/modules/Weaving.ts
@@ -1,4 +1,4 @@
-import {Weaving as CoreWeaving} from 'parser/core/modules/Weaving'
+import {Weaving as CoreWeaving} from 'parser/core/modules/AlwaysBeCasting/Weaving'
 
 export class Weaving extends CoreWeaving {
 	// Pneuma's cure isn't a real action, it can't hurt you
